+++ extensions/openrouter/test-api.ts
@@ -1,2 +1,3 @@
 export { buildOpenRouterImageGenerationProvider } from "./image-generation-provider.js";
 export { openrouterMediaUnderstandingProvider } from "./media-understanding-provider.js";
+export { buildOpenRouterSpeechProvider } from "./speech-provider.js";